+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## Next Release
+- Fix filename issue when saving reports on Windows
+- Add proxy support for `http`, `https`, `socks` and `pac` protocols. Proxy settings can be found under `box configure:settings`
 - Add zip functionality
-- Fix filename issue when saving reports on Windows 
 
 ## 2.5.1 [2020-04-14]